# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Rizzlet
  description: This is a competitive, interactive study tool!
  version: 0.1.0
servers:
  - url: http://localhost:8000/api/
...
paths:
...
  /user/classes:
    get:
      tags:
        - User
      summary: Get all classes for a user
      operationId: userClasses
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
        "422":
          description: Validation exception
        "500":
          description: Server Exception
...
  /class/getDailyGroup:
    post:
      tags:
        - Class
      summary: Get the group for that class for the user logged in
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                classId:
                  type: string
            example:
              classId: "65d679f08f3afb1b89eebfc3"
      operationId: getUserGroup
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  topFour:
                    type: array
                    items:
                      type: object
                      properties:
                        user:
                          type: object
                          properties:
                            id:
                              type: string
                            name:
                              type: string
                        score:
                          type: number
        "422":
          description: Validation exception
        "500":
          description: Server Exception
        "401":
          description: Unauthorized
...